+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * new example i2s_parallel_dimming as an example of using a delay as overall
   brightness control. [#22](https://github.com/liebman/esp-hub75/pull/22)
 
+### Changed
+
+* bump `hub75-framebuffer` to `0.4.0` [#27](https://github.com/liebman/esp-hub75/pull/27)
+
 ## [0.5.0] - 2025-07-16
 
 ### Changed