+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.6.0] - 2025-08-17
+
 ### Added
 
 * added `iram` feature to place the hot rendering function in ram
@@ -119,7 +121,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * initial version
 
 <!-- next-url -->
-[Unreleased]: https://github.com/liebman/esp-hub75/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/liebman/esp-hub75/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/liebman/esp-hub75/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/liebman/esp-hub75/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/liebman/esp-hub75/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/liebman/esp-hub75/compare/v0.2.3...v0.3.0